Refactor RasterRenderer
